isolate parsing to own module
